requirements.txt
irc == 19.0.1
python-dotenv == 0.20.0
requests == 2.22.0